Major codebase reorganization with documentation consolidation and architecture improvements
- Remove 29+ outdated research and investigation docs (9,000+ lines cleanup)
- Add comprehensive TreeView improvement plan with performance analysis
- Update Actor/Relay architecture documentation with refined patterns
- Refactor waveform canvas rendering with improved coordinate handling
- Add unified timeline service architecture with new time types
- Update config system with enhanced reactive triggers
- Consolidate state management patterns and signal handling
- Establish clear technical foundation for future development

+++ shared/src/lib.rs
@@ -1063,9 +1063,9 @@ pub struct WorkspaceSection {
     #[serde(default = "default_timeline_zoom_level")]
     pub timeline_zoom_level: f32,
     #[serde(default)]
-    pub timeline_visible_range_start: Option<f32>,
+    pub timeline_visible_range_start: Option<f64>,
     #[serde(default)]
-    pub timeline_visible_range_end: Option<f32>,
+    pub timeline_visible_range_end: Option<f64>,
 }
 
 impl Default for WorkspaceSection {
